fix: structured the homepage different
