+++ frontend/src/components/NavBar.tsx
@@ -4,17 +4,21 @@ import ProfilePicture from "./ProfilePicture";
 import AppBar from "@mui/material/AppBar";
 import * as React from "react";
 import AppUser from "../model/AppUser";
+import CssBaseline from "@mui/material/CssBaseline";
 
-export default function Navbar({appUser}:{appUser: AppUser}) {
+export default function Navbar({appUser}: { appUser: AppUser }) {
 
-    return(
-        <AppBar position="fixed" sx={{zIndex: (theme) => theme.zIndex.drawer + 1, background: '#2e7d32'}}>
-            <Toolbar sx={{justifyContent: "space-between"}}>
-                <Typography variant="h6" noWrap component="div">
-                    PPC Analytics Tool
-                </Typography>
-                <ProfilePicture appUser={appUser}/>
-            </Toolbar>
-        </AppBar>
+    return (
+        <div>
+            <CssBaseline/>
+            <AppBar position="fixed" sx={{zIndex: (theme) => theme.zIndex.drawer + 1, background: '#2e7d32'}}>
+                <Toolbar sx={{justifyContent: "space-between"}}>
+                    <Typography variant="h6" noWrap component="div">
+                        PPC Analytics Tool
+                    </Typography>
+                    <ProfilePicture appUser={appUser}/>
+                </Toolbar>
+            </AppBar>
+        </div>
     )
 }
